requirements.txt
mock~=4.0.2
moderngl==5.6.1
numpy~=1.19.0
Pillow~=7.1.2
PyGLM==1.0.0